pyproject.toml
[project]
name = "detection-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
